requirements.txt
pandas
numpy
scipy
sklearn
xgboost
statsmodels
matplotlib
scikit-allel[full]
plotly==5.1.0
shap
pomegranate
